[Fix] Answer, ButtonFactory, UI_Logic & UI_Render using them together
- Adds HUDFactory(emtpy) & ButtonPosition Model class;

Keywords: Fix, Answer, ButtonFactory, UI_Logic, UI_Logic, UI_Render
Importance: Medium

+++ core/src/com/mygdx/game/UI/UI_Logic.java
@@ -8,7 +8,8 @@ import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 
 public class UI_Logic {
 
-    public TextButton createButton(String txt, int[] coords, ChangeListener changeListener){
+    public TextButton createButton(String txt, float[] coords,
+                                   ChangeListener changeListener){
         BitmapFont font = new BitmapFont();
         Skin skin = new Skin(); // used for texturing the button's background
         TextButtonStyle textButtonStyle = new TextButtonStyle();
